+++ galaxy2_save_data/src/chunks/game/galaxy.rs
@@ -4,7 +4,7 @@ use bilge::prelude::*;
 use binrw::binrw;
 use galaxy_save_core::{
     bin::{BinaryDataContentHeaderSerializer, Chunk, HeaderSerializer},
-    hash::HashCode,
+    hash::{HashCode, HashCode16},
 };
 
 #[cfg(feature = "serde")]
@@ -37,7 +37,7 @@ pub struct SaveDataStorageGalaxy {
 impl SaveDataStorageGalaxy {
     /// Returns a reference to the [`SaveDataStorageGalaxyStage`] corresponding to the key.
     pub fn get(&self, galaxy_name: impl Into<HashCode>) -> Option<&SaveDataStorageGalaxyStage> {
-        let galaxy_name = galaxy_name.into().into_raw() as u16;
+        let galaxy_name = HashCode16::from(galaxy_name.into());
 
         self.galaxy.iter().find(|v| v.galaxy_name == galaxy_name)
     }
@@ -47,7 +47,7 @@ impl SaveDataStorageGalaxy {
         &mut self,
         galaxy_name: impl Into<HashCode>,
     ) -> Option<&mut SaveDataStorageGalaxyStage> {
-        let galaxy_name = galaxy_name.into().into_raw() as u16;
+        let galaxy_name = HashCode16::from(galaxy_name.into());
 
         self.galaxy
             .iter_mut()
@@ -72,7 +72,7 @@ impl Chunk for SaveDataStorageGalaxy {
 pub struct SaveDataStorageGalaxyStage {
     /// The hashed internal name of the galaxy, truncated to the least significant 16 bits.
     #[header_serializer(name = "mGalaxyName")]
-    galaxy_name: u16,
+    galaxy_name: HashCode16,
 
     /// The size of the serialized struct, in bytes.
     #[header_serializer(name = "mDataSize")]