+++ galaxy2_save_data_json/src/cli.rs
@@ -10,6 +10,14 @@ pub struct Args {
     /// The output save or JSON file path.
     pub output: Option<String>,
 
+    /// The newline-separated hash labels file path.
+    #[arg(short, long)]
+    pub labels: Option<String>,
+
+    /// Reject hash labels not found in the labels file.
+    #[arg(short, long, requires("labels"))]
+    pub strict: bool,
+
     /// Parse the save file even if the header is invalid.
     #[arg(short, long)]
     pub force: bool,