+++ galaxy_save_data_json/src/cli.rs
@@ -10,6 +10,14 @@ pub struct Args {
     /// The output save or JSON file path.
     pub output: Option<String>,
 
+    /// The newline-separated hash labels file path.
+    #[arg(short, long)]
+    pub labels: Option<String>,
+
+    /// Reject hash labels not found in the labels file.
+    #[arg(short, long, requires("labels"))]
+    pub strict: bool,
+
     /// The source or target console.
     #[arg(short, long, default_value_t, value_enum)]
     pub platform: Platform,
@@ -31,4 +39,23 @@ pub enum Platform {
 
     /// Nintendo Switch (Super Mario 3D All-Stars)
     Switch,
+}
+
+/// A character encoding for labels prior to hashing.
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum LabelEncoding {
+    /// Shift JIS
+    ShiftJis,
+
+    /// UTF-8
+    Utf8,
+}
+
+impl From<Platform> for LabelEncoding {
+    fn from(platform: Platform) -> Self {
+        match platform {
+            Platform::Wii | Platform::ShieldTv => Self::ShiftJis,
+            Platform::Switch => Self::Utf8,
+        }
+    }
 }